test(canon): split grown test files to satisfy source-length gate
